requirements.txt
langgraph==0.3.18
langchain-core>=0.1.0
pydantic>=2.7.4
pyyaml>=5.3
flask>=2.0.0
flask-cors>=3.0.0